requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0 
drf-yasg>=1.21.7 